requirements.txt
aioschedule==0.5.2
asyncpg==0.26.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-auth==2.11.0
google-auth-oauthlib==0.5.2
greenlet==1.1.3
gspread==5.5.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
psycopg==3.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
